+++ src/hooks/useAsync.js
@@ -1,0 +1,20 @@
+import { useState } from "react";
+
+export function useAsync(callback) {
+  const [pending, setPending] = useState(false);
+  const [error, setError] = useState(null);
+
+  const wrappedFunction = async (...args) => {
+    try {
+      setPending(true);
+      setError(null);
+      return await callback(...args);
+    } catch (error) {
+      setError(error);
+    } finally {
+      setPending(false);
+    }
+  };
+
+  return [pending, error, wrappedFunction];
+}
